Add azure package for share azure constants and functions
