+++ lib/output/writer/azure_queue_storage.go
@@ -4,11 +4,10 @@ package writer
 
 import (
 	"context"
-	"errors"
 	"fmt"
-	"net/url"
-	"strings"
 	"time"
+
+	"github.com/Jeffail/benthos/v3/internal/service/azure"
 
 	"github.com/Azure/azure-storage-queue-go/azqueue"
 	"github.com/Jeffail/benthos/v3/internal/bloblang"
@@ -31,89 +30,21 @@ type AzureQueueStorage struct {
 	stats metrics.Type
 }
 
-const (
-	azQueueEndpointExp  = "https://%s.queue.core.windows.net"
-	devQueueEndpointExp = "http://localhost:10001/%s"
-	azAccountName       = "accountname"
-	azAccountKey        = "accountkey"
-	devAccountName      = "devstoreaccount1"
-	devAccountKey       = "Eby8vdM02xNOcqFlqUwJPLlmEtlCDXJ1OUzFT50uSRZ6IFsuFq2UVErCz4I6tq/K1SZFPTOtr/KBHBeksoGMGw=="
-)
-
-// parseConnectionString extracts the credentials from the connection string.
-func parseConnectionString(input string) (storageAccount, storageAccessKey string, err error) {
-	// build a map of connection string key/value pairs
-	parts := map[string]string{}
-	for _, pair := range strings.Split(input, ";") {
-		if pair == "" {
-			continue
-		}
-		equalDex := strings.IndexByte(pair, '=')
-		if equalDex <= 0 {
-			fmt.Println(fmt.Errorf("invalid connection segment %q", pair))
-		}
-		value := strings.TrimSpace(pair[equalDex+1:])
-		key := strings.TrimSpace(strings.ToLower(pair[:equalDex]))
-		parts[key] = value
-	}
-	accountName, ok := parts[azAccountName]
-	if !ok {
-		return "", "", errors.New("invalid connection string")
-	}
-	accountKey, ok := parts[azAccountKey]
-	if !ok {
-		return "", "", errors.New("invalid connection string")
-	}
-	return accountName, accountKey, nil
-}
-
 // NewAzureQueueStorage creates a new Azure Queue Storage writer type.
 func NewAzureQueueStorage(
 	conf AzureQueueStorageConfig,
 	log log.Modular,
 	stats metrics.Type,
 ) (*AzureQueueStorage, error) {
-	var err error
-	if len(conf.StorageAccount) == 0 && len(conf.StorageConnectionString) == 0 {
-		return nil, errors.New("invalid azure storage account credentials")
+	serviceURL, err := azure.GetQueueServiceURL(conf.StorageAccount, conf.StorageAccessKey, conf.StorageConnectionString)
+	if err != nil {
+		return nil, err
 	}
-	var storageAccount string
-	var storageAccessKey string
-	var endpointExp = azQueueEndpointExp
-	if len(conf.StorageConnectionString) > 0 {
-		storageAccount, storageAccessKey, err = parseConnectionString(conf.StorageConnectionString)
-		if err != nil {
-			return nil, err
-		}
-		if strings.Contains(conf.StorageConnectionString, "UseDevelopmentStorage=true;") {
-			storageAccount = devAccountName
-			storageAccessKey = devAccountKey
-			endpointExp = devQueueEndpointExp
-		} else if strings.Contains(conf.StorageConnectionString, devAccountName) {
-			endpointExp = devQueueEndpointExp
-		}
-	} else {
-		storageAccount = conf.StorageAccount
-		storageAccessKey = conf.StorageAccessKey
-	}
-	if len(storageAccount) == 0 {
-		return nil, fmt.Errorf("invalid azure storage account credentials: %v", err)
-	}
-	var credential azqueue.Credential
-	if len(storageAccessKey) > 0 {
-		credential, _ = azqueue.NewSharedKeyCredential(storageAccount, storageAccessKey)
-	} else {
-		credential = azqueue.NewAnonymousCredential()
-	}
-	p := azqueue.NewPipeline(credential, azqueue.PipelineOptions{})
-	endpoint, _ := url.Parse(fmt.Sprintf(endpointExp, storageAccount))
-	serviceURL := azqueue.NewServiceURL(*endpoint, p)
-
 	s := &AzureQueueStorage{
 		conf:       conf,
 		log:        log,
 		stats:      stats,
-		serviceURL: &serviceURL,
+		serviceURL: serviceURL,
 	}
 
 	if s.ttl, err = bloblang.NewField(conf.TTL); err != nil {
